refactor(geo_core): Move sphere_metrics to spherical shape implementations
- Move sphere distance calculation functions to SphericalSolid3D/SphericalSurface3D methods
- Remove geo_core/sphere_metrics.rs (improper dependency direction)
- Add distance_to_line_segment, distance_to_ray, distance_to_infinite_line methods to both shapes
- Update collision implementations to use new instance methods
- Fix architecture: geo_primitives should not depend on geo_core helpers for shape-specific logic
- All tests passing (24 spherical tests)
